requirements.txt
absl-py==2.1.0
anyio==4.4.0
astunparse==1.6.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
efficientnet-pytorch==0.7.1
exceptiongroup==1.2.2
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.4
h11==0.14.0
h5py==3.11.0
hnswlib==0.8.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.35.0
itsdangerous==2.2.0
Jinja2==3.1.4
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
pillow==10.4.0
protobuf==4.25.4
Pygments==2.18.0
python-dotenv==1.0.1
python-telegram-bot==21.4
requests==2.32.3
rich==13.7.1
scikit-image==0.24.0
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
sympy==1.13.2
termcolor==2.4.0
tifffile==2024.8.10
torch==2.4.0
torchvision==0.19.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0